♻️(frontend) Migrate current user state to Pinia store
