+++ src/web/presentation/frontend/package.json
@@ -23,6 +23,7 @@
     "@sentry/vue": "^10.58.0",
     "@tanstack/vue-table": "^8.21.3",
     "openapi-fetch": "^0.17.0",
+    "pinia": "^3.0.4",
     "reka-ui": "^2.9.10",
     "vue": "^3.5.38",
     "vue-component-meta": "^3.3.5",
