+++ src/web/presentation/frontend/src/stores/currentUser.test.ts
@@ -1,5 +1,6 @@
+import { createPinia, setActivePinia } from 'pinia'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
-import { useCurrentUser } from './useCurrentUser'
+import { useCurrentUserStore } from './currentUser'
 
 const mockGetMe = vi.fn()
 
@@ -7,8 +8,9 @@ vi.mock('@/api/utilisateur', () => ({
   getMe: () => mockGetMe(),
 }))
 
-describe('useCurrentUser', () => {
+describe('useCurrentUserStore', () => {
   beforeEach(() => {
+    setActivePinia(createPinia())
     vi.clearAllMocks()
   })
 
@@ -20,41 +22,41 @@ describe('useCurrentUser', () => {
     const userData = { email: 'test@example.com', prenom: 'Jean', nom: 'Dupont' }
     mockGetMe.mockResolvedValue(userData)
 
-    const { user, displayName, loading, fetch } = useCurrentUser()
+    const store = useCurrentUserStore()
 
-    expect(user.value).toBeNull()
-    expect(loading.value).toBe(false)
+    expect(store.user).toBeNull()
+    expect(store.loading).toBe(false)
 
-    const promise = fetch()
-    expect(loading.value).toBe(true)
+    const promise = store.fetch()
+    expect(store.loading).toBe(true)
 
     await promise
 
-    expect(loading.value).toBe(false)
-    expect(user.value).toEqual(userData)
-    expect(displayName.value).toBe('Jean Dupont')
+    expect(store.loading).toBe(false)
+    expect(store.user).toEqual(userData)
+    expect(store.displayName).toBe('Jean Dupont')
   })
 
   it('handles fetch error gracefully', async () => {
     const networkError = new Error('Network error')
     mockGetMe.mockRejectedValue(networkError)
 
-    const { user, error, fetch } = useCurrentUser()
+    const store = useCurrentUserStore()
 
-    await fetch()
+    await store.fetch()
 
-    expect(user.value).toBeNull()
-    expect(error.value).toBe(networkError)
+    expect(store.user).toBeNull()
+    expect(store.error).toBe(networkError)
   })
 
   it('deduplicates concurrent fetch calls', async () => {
     const userData = { email: 'test@example.com', prenom: 'Jean', nom: 'Dupont' }
     mockGetMe.mockResolvedValue(userData)
 
-    const { fetch } = useCurrentUser()
+    const store = useCurrentUserStore()
 
-    const promise1 = fetch()
-    const promise2 = fetch()
+    const promise1 = store.fetch()
+    const promise2 = store.fetch()
 
     await Promise.all([promise1, promise2])
