+++ src/web/presentation/frontend/vite.config.ts
@@ -35,7 +35,7 @@ export default defineConfig(({ command }) => ({
     environment: 'jsdom',
     coverage: {
       provider: 'v8',
-      include: ['src/utils/**', 'src/api/**', 'src/composables/**', 'src/features/**/composables/**'],
+      include: ['src/utils/**', 'src/api/**', 'src/stores/**', 'src/composables/**', 'src/features/**/composables/**'],
       exclude: ['**/*.stories.ts', 'src/composables/dnd/**'],
       thresholds: {
         'src/utils/**': { statements: 80, branches: 80, functions: 80, lines: 80 },
